+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Additional dynamic data available in Page Title
+
 ## [0.6.1] - 2021-02-01
 
 ### Fixed